refactor: portal routing, split portal into ticketing & kb pages
 - kb pages does not required user login to access
 - ticketing pages (ticket list/new ticket) require user login
