+++ desk/src/pages/portal/kb/KnowledgeBase.vue
